Fix crash related to position update timer in MFPlayerSession
The MFPlayerSession instance can outlive its associated MFPlayerControl
instance, since MFPlayerSession is a COM object that can be referenced by
other objects and will only be deleted when its reference count goes to 0.
So, it could end with a dangling pointer to the released MFPlayerControl
instance, which would cause a crash when it is used in the timer callback.
This patch should help preventing crashes in other conditions as well.

Fixes: QTBUG-97069
Change-Id: Id1e303e9727019d00a3765dcde35805c51ca335f
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit 4e646c2586c27183497eb8078829a4000c06edcd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>