Add test for change application state while recording
This test tests if the a media recording can finish successfully when
the application state changes to a inative state while recording. This
can happen in mobile devices when you are recording and user changes
the app or if the user gets a popup from the OS like a permission popup.

Change-Id: I4a4dbca3e47b4aa345be4ea3fa257eac523bb8a1
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit edf911e3c162f161ce553d67cfd7f7950f9b435c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>