Add test for change application state while recording
This test tests if the a media recording can finish successfully when
the application state changes to a inative state while recording. This
can happen in mobile devices when you are recording and user changes
the app or if the user gets a popup from the OS like a permission popup.

Pick-to: 6.2
Change-Id: I4a4dbca3e47b4aa345be4ea3fa257eac523bb8a1
Reviewed-by: Lars Knoll <lars.knoll@qt.io>