macOS/iOS: Query for camera and microphone authorization status
In order to gracefully handle cases where permission to use camera or
microphone is not granted, query for authorization status beforehand
and prompt for permission when the status is not determined.
The API to prompt for permission is asynchronous so make sure to
set up the AVCaptureInputs only once permission is granted.

Task-number: QTBUG-97408
Pick-to: 6.2
Change-Id: Ifeb94003b83626483336e0b8c69cb40e5841edd8
Reviewed-by: Lars Knoll <lars.knoll@qt.io>