Implement QAndroidImageCapture::captureToBuffer()
Implemented the captureToBuffer based on the current implementation.
Found that the image was not created on the right orientation. Fixed
that by correcting the orientation after the picture was taken.
Small refactor on cameraandroidsession because of different arguments.

Fixes: QTBUG-97587
Change-Id: Ieb0c07f3b397df789f3f43d033ced76c01e8f33f
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 60f9c575c2e21f5dfc08386e537ea56d5759b826)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>