Refactor Windows QAudioDecoder to conform to tst_qaudiodecoderbackend
This is a cleanup of the QAudioDecoder backend for Windows. The patch
enables the decoder to work without the resampler, if it could not be
instantiated. It adds warning messages and error signals to notify
a user about issues. The decoder passes tst_qaudiodecoderbackend with
minor adjustments that will follow.

Fix QWindowsIUPointer self copy operation.

Pick-to: 6.2
Change-Id: Ie962ac7615b957746f85dcc4ec49af38bbda2373
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>