Enable QAudioDecoder backend tests for Windows
The patch enables the QAudioDecoder backend tests for Windows with minor
adjustments. The expected duration time of the resampled file has been
reduced as the Windows resampler produces slightly shorter audio after
resampling. Additionally, the expected error code when reading invalid
file is changed to ResourceError, which is better suited for this case
than AccessDeniedError that relates to lack of permission.

Pick-to: 6.2
Change-Id: Iaf905a66c6e03d6ba6d5c9afd4fc8e88b9cb7840
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Doris Verria <doris.verria@qt.io>