Fix QVideoWidget::setFullScreen()
setFullScreen(true) would always cause the video widget to be shown
on the primary screen instead of the current screen the app is being
shown on.

The reason for this is that setting the Qt::Window flag on the widget
will cause its position to be interpreted as relative to the screen
instead of its parent. This position is usually on the first screen,
so showFullScreen() will show it there.

Fix this by moving the video widget to the physical position on the
screen that it had inside its widget hierarchy before calling
showFullScreen(). Also restore that position on setFullScreen(false),
so that its position is correctly restored even without layouts.

Fixes: QTBUG-97895
Change-Id: I5207afb276e9369938d1c3d8814d9878367c6895
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit d6cef755f6874190d144d4d574c3c0d78f38dec4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>