Use system clock on the pipeline instead of audio device clock
By default the pipeline uses the audio output device clock.
When changing the audio output, the clock is lost and the pipeline
must be paused and then put to the play state again. However, this
process results in a complete stop of the pipeline. To avoid the clock
switching problem, we can enforce the system clock on the pipeline
that will never be changed.

Change-Id: I00542060d14aa546f5945d863c59831caacf9fdc
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 0e259b5a0b90544d706594e581befaedcf914bdb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>