Disconnect outputs from input selector when no track is selected
The current implementation connects nullTrack to an output via input
selector when disabling a track. However, this solution does not work
well with other gstreamer elements. For example, when a video track is
disabled, all the events coming from the video sink do not arrive at
the decoder as they are routed to the nullTrack and vanish there. This
results in unstable and unpredictable player behavior.

This patch simply disconnects the input selector from the output when
the application disables a track.

Pick-to: 6.2
Change-Id: I2eccfa637c92e716e1716edd92a85a6f66ae2b28
Reviewed-by: Lars Knoll <lars.knoll@qt.io>