Windows: Fix crash while reloading QMediaPlayer
MFStream uses an event for reading data that used a pointer to an
externally owned object that due to a race-condition could
dereference the pointer after the pointed object had already been
deleted, resulting in a crash. This patch changes it to use a weak
reference and test if the object is still valid.

Fixes: QTBUG-97992
Change-Id: I4ef37ad49036f9fa33e5f4f753c07c37398996c0
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 2e857dd4b2d0b7564422b2d99ae8cad77c10cc5d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>