Change Android MediaPlayer SeekTo to the Closest Frame
The default behavior for seekTo in Android is to seek the previous Sync
Frame (the previous I-frame), disregarding the provided time. Depending
on the video file, this can be a 5-second difference.
This patch will change this behavior for the closest frame of the
provided time. It is still not the exact time, but it will be closer
than before. This change is only possible for Android API > 26.
Below that will continue to seek to the previous Sync Frame.

Fixes: QTBUG-97415
Change-Id: I9e698b023d30c0e527180198970430f39fda1f93
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Pekka Gehör <pekka.gehor@qt.io>
(cherry picked from commit 0c1bb90691780703eabc7ed1954d6b49fe5da580)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>