Fix 'devices' example to build with MinGW
src/gui/kernel/qwindowdefs.h #defines
main to qMain and qtentrypoint_win.cpp expects qMain
to have 2 parameters in its signature: argc and argv.

The 'devices' example was missing the parameters, which caused the app
to fail at link time with

  qtentrypoint_win.cpp:97: undefined reference to `qMain(int, char**)'

Add the parameters to the example signature.

Pick-to: 6.2
Task-number: COIN-762
Change-Id: I8baa4e84abf4c08f462f041a30d261023282a0b9
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>