FFmpeg: support for seeking
FFmpeg usually seeks on dts, not pts and will always sink to a
key frame. This means we do not get to the exact position that is
requested. To have frame exact seeking, some more work is required.

After seeking to the requested timestamp, check the pts of next frame.
If it's behind our position seek to an earlier timetamp and repeat.
Then signal the decoders that we seek to a new position and single
step them until we reach it.

Change-Id: I00edade0c713bb8b2bd4ce716211d0d6439e3da8
Reviewed-by: Lars Knoll <lars.knoll@qt.io>