Use Core Audio API for playing sounds to get better control over playout
This patch replaces the waveOut interface with the IAudioClient
interface that provides better control over the audio output. waveOut
introduced unpredictable delays in the playout described in
QTBUG-97861. IAudioClient is lower level interface than waveOut.
It requires audio resampling to match Windows audio mixer device,
which is done with QWindowsAudioResampler.

Fixes: QTBUG-97861
Pick-to: 6.2
Change-Id: I6f0fcfdc0d43257e15b84f7527600f97c26a0306
Reviewed-by: Lars Knoll <lars.knoll@qt.io>