FFmpeg: Clean up the thread looping and blocking
Each thread should really only block on one place and on it's own
waitcondition, otherwise we can easily end up with dead locks.

To achieve this move the checks on which we would block into the
virtual shouldWait() method. Also wake the previous thread in the
pipeline when we consume a packet/frame.

Change-Id: I4aa0e69c3e943efb1c202411c1e117ed30986f5a
Reviewed-by: Lars Knoll <lars.knoll@qt.io>