Fix orientation on Android QML VideoOutput on landscape
In Android, recent changes in QAndroidCameraSession made the preview
orient itself automatically. On a rotation to landscape, the
VideoOutput QML component is remade with different widths and heights
and scales the content assuming the preview is not oriented.
I altered to transpose the scale when in landscape mode on Android.

Fixes: QTBUG-97080
Change-Id: Ic35ec4e99ab1f15fa8d4db9cd0d7503158bef053
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 7337564717a1bb2f514b8eba78a206597dbe533f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>