Set Camera active on a new camera session is added
In some code path with QML components it was possible for the setActive
of the camera to fail because the it did not have a session at the time
That would happen even if a session is added later. So this patch sets
the camera active when a new media camera session is added.

Fixes: QTBUG-98559
Change-Id: I0aee0dccdb0e41f393264cd786ea7c24c0e4ff69
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 5cc076819278a1a039dc3fe8358535d01591539a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>