AVFMediaAssetWriter: Stop writing to file in case of error
When recording, we didn't check if the AVAsssetWriter has failed (for
various reasons), but continued to capture and tried to write the
buffers. The asset writer couldn't finish writing either if it was in a
failed state, making the recording impossible to stop.

To fix, in case of an internal error, stop the recording and make sure
to set the failure reason to the error message.

Fixes: QTBUG-98262
Change-Id: I2daccdb711f9a310d7939954064ae1152fc6c18a
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
(cherry picked from commit f835e2871cac9473add920b3fb69e20524152656)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>