Fix OpenGL-based camera preview rendering on macOS
Change f117648a2d4f443c2bb90ff27070ae46d97a4297 enabled OpenGL-based
video on macOS. Do the same for the camera capture output. This will
currently be functional only if the pixel format is of the type
kCVPixelFormatType_32BGRA.

We set the pixel format of the viewfinder to the pixel format we set to
the camera. However, if the rhi backend is OpenGL, we need to force the
kCVPixelFormatType_32BGRA pixel format on it. For this reason, we need
to update the pixel format after we have set the rhi to the video
sink.

Try to unify the code between AVFCameraRenderer and AVFVideoRenderer
as the pixel buffer settings are similar between the two.

Fixes: QTBUG-98860
Change-Id: I8a16e82257b24d81b13dabcad7678b543bf78eac
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 2f4c163aae2b08eae27d302f7c535f37bd349509)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>