Remove the AVPlayerItemVideoOutput from player item when changing media
When updating the player item of the AVPlayer, make sure to remove
the video output (AVPlayerItemVideoOutput) from it. It will be
added again to the updated player item on the next call to
copyPixelBufferFromLayer.
This fixes issues where the last frame of the previous video source
was sometimes flashed before rendering the new one.

Fixes: QTBUG-87000
Change-Id: Iec66f0e27efe621d1992a2a0f9f5060aa51f7076
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>