Fix qmediaplayerbackend loadMediaInLoadingState test in Android
This particular test expected for a media status of loading after a
setSource. On Android that only happened on the first time that a source
was set. That was changed to properly change to state to a loadingMedia
when the data source will be changed.

Fixes: QTBUG-99181
Change-Id: Ic0e1d9e9517edffd9b59d098c9ff70df0673c9c1
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 33e1b6d61d2af3247e224d1cd08f6f0e051f99d9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>