Multimedia examples: Add camera and audio input entitlements
Apple needs apps to specify proper entitlements in order to grant the
executable permissions to use services like camera and microphone.

The multimedia examples using microphone and camera were sometimes
crashing because of missing entitlements. To fix, add an entitlement
file to all examples using these services, and add a post-build
command to codesign the executables with the specified entitlements.

Task-number: QTBUG-98419
Pick-to: 6.3 6.2
Change-Id: I6a578def1f6a41b4d106ee49c0c32fad304ef3cb
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>