Windows: Get list of supported codecs from Media Foundation
This patch uses a table for container compatibility and queries Media
Foundation for the actually supported audio and video codecs, instead
of using hardcoded definitions for everything.

Also, QImageWriter is queried for supported image formats, and the
container selection for audio-only files has been fixed.

It is now possible to encode audio as flac, alac, etc.

Pick-to: 6.2 6.3
Change-Id: Ie8fcbf6582ff1785ac0c77def8cebb0aef100adb
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Piotr Srebrny <piotr.srebrny@qt.io>