AVFMediaAssetWriter: Properly reset queues and writer inputs before starting new writing
We were resetting the video queue only if there was a video input
attached to the session, causing the old video queue to still be valid
otherwise. This led to crashes when the next recording would be audio-
only because of a failed assert.
To fix, release the audio and video queues and set them to null when
we set up the writer for a new recording. Do the same for the camera
and audio writer input.

Also, create an audio queue only if there is an audio input attached
to the session.

Fixes: QTBUG-99296
Fixes: QTBUG-99176
Change-Id: If08b7f514ba3f264509ee78b494d8d1544991ad6
Reviewed-by: Piotr Srebrny <piotr.srebrny@qt.io>
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
(cherry picked from commit 4bae921b03c00714efb3ef1eea06d4ca8f9b70d4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>