Fix SurfaceText test in Android
Looking at logcat, it was possible to understand that the frames were
parsed, but they were not updated. Looks like that in android it is
necessary to call the frame map for the AndroidTextureVideoBuffer to
updateReadbackFrame. Updated the example to force that.

Fixes: QTBUG-99358
Change-Id: Id8a0e622158f836813b7ed266ad3580aff68171f
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 203cb0f9523739c2c3583abd69e55fa49c41588e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>