Reintroduce a plugin architecture for multimedia backends
This is required, so we can introduce FFMpeg as a technology preview,
and have both backends available in one compiled version of Qt.

The plugin archicture is much simpler than in Qt 5, using one plugin
per backend, not multiple.

Export a couple of classes and methods from Qt Multimedia that are
required for the backends to work.

Change-Id: I99ab324c5ee018a3cefcf32d82f16a0bf6b5d409
Reviewed-by: Piotr Srebrny <piotr.srebrny@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>