FFmpeg: Add more generic clock handling
Add Clock and ClockController classes to abstract away
clock handling, and be able to deal with multiple concurrent
clocks that need to be synced against each other.

This works nicely if the audio clock is the master, but will
need some additional work to ensure synchronization when
clocks diverge and for the case where audio is not the master
clock (for playback that's basically only for live streams).

Change-Id: I403bd222b42ce666453d27ba8a212c12fc90a90b
Reviewed-by: Lars Knoll <lars.knoll@qt.io>