Use PulseAudio directly for low-level audio processing
It turns out that using gstreamer for QAudioSink/Source is introducing
a very large latency for low-level audio (in the order of 3-400 ms).

For low-latency use cases that is not acceptable, so use pulseaudio
directly for QAudioSource/Sink if cmake finds it. When not found,
fall back to the old code path that uses gstreamer.

This required some slight changes to QAudioInput/Output as well, as
we now get our device ids from pulseaudio. To make it work, simply
use the gstreamer pulsesrc/pulsesink elements directly instead of
going through the gst_device_* APIs.

Also disable the FFmpeg backend on Linux without pulseaudio, as
we need it for audio IO.

Change-Id: I54d26400773626e7c1fed9312927cedfe0facbc0
Reviewed-by: Piotr Srebrny <piotr.srebrny@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>