FFmpeg: Fix uneven video frame timings
Fix the timeout calculcations for when to render the next video frame.
We've been calculating the sleep time using the VideoRenderer clock.
This is slightly wrong, as it's not always 100% in sync with the audio
clock.

This wasn't very visible on Unix systems, where timers and timeouts on
on the event loop are rather precise, but caused noticable issues on
Windows.

Change-Id: Ic6b43a8b38e2fa612a9456ba5e344dc33fa4c373
Reviewed-by: Piotr Srebrny <piotr.srebrny@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>