Use clockwise rotation for the QVideoFrame::rotationAngle()
The counterclockwise rotation of rotationAngle() does not play well
with the multmedia backends. The WMF MFVideoRotationFormat that
describes the counterclockwise rotation of the picture, must be applied
as clockwise rotation to render frame the correct way.

This patch changes the interpretation of the rotationAngle() to be
clockwise rotation. This fixes the rendering for WMF and gstreamer.
AVFoundation is updated to account for the direction change.

Fixes: QTBUG-100337
Change-Id: I284367629dad27fea95da52af4ca2d71d9412b09
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 099e9a697b92e02666f308c1789c685990e45206)
Reviewed-by: André de la Rocha <andre.rocha@qt.io>