Do not access NV_DX_interop functions out of OpenGL context
An attempt to use NV_DX_interop functions out of OpenGL context results
in an exception. Thus, we do not unregister the D3D9 texture from
the OpenGL context when it is gone.

This patch also clears m_glHandle variable after it has been unregister
in some rare case when we cannot lock it.

Pick-to: 6.2 6.3
Fixes: QTBUG-100665
Change-Id: Ie4bc02dbaccec12d92f8e274eeb420482a5f2d58
Reviewed-by: André de la Rocha <andre.rocha@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>