FFmpeg: Get the audio decoder working
This gets the audio decoder working with the FFmpeg backend.

The example still has some issues with -p and the fact that FFmpeg
often decodes using float based samples (which our QWaveDecoder
doesn't like). But those issues seem to be related to the example
and QWaveDecoder rather than the audio decoder backend.

Change-Id: I1f9d21bb80f96fb3da55a2f7795153b041dba2c4
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Piotr Srebrny <piotr.srebrny@qt.io>