Restructure the HW codec support
Disentangle HW codec support from the conversion into RHI compatible
textures. Move the latter to a new TextureConverter class that controlled
by the video sink, and only set on the video buffer when it passes through
the sink.

Properly create a HW context for the camera and add some preliminary
support to the encoder on macOS.

Adjust the VAAPI code to the new structure.

Change-Id: I21ecf67fd276dfaa3ee7e04c80fe2304a3d536be
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Piotr Srebrny <piotr.srebrny@qt.io>