Add internal ability to adjust RTSP latency
When using the gstreamer backend, the RTSP latency is set to 2000ms by
default. This latency creates issues for cases where near realtime
camera feedback is critical for directing the user.

This change adds the ability to turn 3 knobs in the RTSP jitter buffer
pipeline using the following environment variables:
    1. QT_MEDIA_RTSP_LATENCY
        An unsigned integer value expressed in ms. This dictates the
        high water mark of the latency buffer.
        Default: 2000 ms
    2. QT_MEDIA_RTSP_DROP_ON_LATENCY
        May be set to 1 for on or 0 for off
        When turned on, the jitter buffer will drop packets that arrive
        after the high water mark of the buffer has been filled. No new
        packets will be accepted until the high water mark recedes
        sufficiently to allow new packets to be stored in the buffer.
        Default: off
    3. QT_MEDIA_RTSP_DO_RETRANSMISSION
        May be set to 1 for on or 0 for off
        When set to true, the jitter buffer will request a re-send of
        dropped packets.
        Default: off

This change is strictly for a customer project, but may be helpful in
understanding what additions to the public API would be helpful in the
future.

Pick-to: 6.2 6.4 6.5
Change-Id: Icede93161d72936970f4835c6443e369789b4c38
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Doris Verria <doris.verria@qt.io>