Fix android media plugin issues
Android Multimedia plugin was not able to create OpenGL context after
changing to plugin because the plugin changes the thread where the
plugin runs. This is due to being loaded at runtime. This patch fixes this
by changing the thread to the OpenGL context thread if needed.
Also fixes a missing video output start that prevented the camera
preview to work.

Fixes: QTBUG-100773
Pick-to: 6.3
Change-Id: I8bc2a3e87dedf2c54d20904afb8a68fdff2420ca
Reviewed-by: Pekka Gehör <pekka.gehor@qt.io>
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>