Windows: Fix lack of precision in positioning while playing flac
When changing the position while playing a flac file, the actually played
position is not the one specified precisely; it seems to snap to some
low precision point. However, the precise position is being specified
correctly at the WMF level, and flac files not being handled differently
in any way. This behavior seems to be caused by an issue in WMF. However,
it seems that resetting the audio output routing when playing the file for
the first time avoids the issue without causing side effects.

Fixes: QTBUG-100181
Pick-to: 6.3 6.2
Change-Id: Ifbf7770df0040231dc1b75edd9b2b34d6c870ef0
Reviewed-by: Piotr Srebrny <piotr.srebrny@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>