AVFCameraSession - don't use preview layer
It's not in use essentially and was a legacy when we, indeed, were
adding it as a sub-layer in a view. Starting from macOS Monterey,
creating such a layer using +layerWithSession: results in a broken
AVCaptureSession - frames/sample buffers stop coming after the first
25 frames. Also, qt application may break other non-qt apps,
having their own capture session/viewfinders frozen too.

Fixes: QTBUG-99641
Pick-to: 6.3 6.2
Change-Id: I75dfba5db0d50bd015fcd6904a0c5c8ea800230f
Reviewed-by: Doris Verria <doris.verria@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>