Simplify the clock handling
The ClockController is now the only place having a timer. Clock
instances basically retrieve their time from the controller, and
can tell it where they are through the timeUpdated() method.

This method will now always return the current time where one should be.
If the Clock is being used as the master (ie. it controls the timing
of the pipeline), the returned value will be the same as the input
value.

In other cases, it might return a value different from where the Clock
currently is. This can then be used by the Clock to adjust itself to
those changes.

Change-Id: I4d2e5952d388f200d4a80f6d65c15c331912112f
Reviewed-by: Rafael Roquetto <rafael.roquetto@qt.io>