Fix missing main.qml in examples
These examples use resources files with qml files inside but the
resource files were added in the project sources. That prevented
the app to find and use the resource files. To fix the issue the
resources were added using qt6_add_resources macro.
Note that adding the cmake flag set(CMAKE_AUTORCC ON) would also
work, but the qt6_add_resources macro is more consistent with the
rest of the examples.

Fixes: QTBUG-101518
Pick-to: 6.3
Change-Id: I5fb2481d5576683171a2292bdc35e417ee6d8fab
Reviewed-by: Piotr Srebrny <piotr.srebrny@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>