QNX: fix lib linkage
We shouldn't be linking against libmmrndcore - this library is used by
the mm-renderer resource manager, and is not meant to be used by client
applications.

libmmrndcore exposes several symbols that clash which symbols exposed by
libmmrndclient (e.g. mmr_connect), but that provides a "server-side"
implementation.

libstrm needs to be brought along, as it was being implicitly pulled by
libmmrndcore, and is still needed by libmmrndclient.

Pick-to: 6.3
Change-Id: Ie2dacc387b7be8d8d19d6ede05a6cd0661a20e7c
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: James McDonnell <jmcdonnell@blackberry.com>