Windows: Rework workaround for WMF issue
What seems to be a WMF issue was causing a crash with specific videos
and also issues when changing position of flac files. A previous
workaround for it was depending on the way the media playing was started
and thus seemed to fix the issue when tested with the Widgets-based
Media Player example, but still failed with the QML-based one. This
patch seems to fix it for all cases.

Fixes: QTBUG-100854
Fixes: QTBUG-100181
Pick-to: 6.2 6.3
Change-Id: I7e46d72b5e656383442d2866ef93f3183c9f9fa7
Reviewed-by: Lars Knoll <lars.knoll@qt.io>