Provide QRhiTextures directly from QVideoFrame to the renderers
This patch adds a method to provide QRhiTexture directly from
QVideoFrame instead of an opaque handle. This will be use in
the followup patches to replace the handle method.

The ownership of textures in QVideoTextureHelper::updateRhiTexture is
now expressed with unique_ptr and the function operates on individual
planes which improves composability.

Additionally, the patch fixes a problem when reusing textures in mapped
mode, where the reused texture retained incorrect pixel format. Now,
the pixel format is updated.

Change-Id: I2381a5dcbabfae3c27c76c724d88bcf5d9bbb23d
Reviewed-by: Lars Knoll <lars.knoll@qt.io>