Fix extremely slow conversion of video frames to QImage on QNX
mapping a QVideoFrame on QNX maps an uncached memory buffer. This means
that the old code reading data byte for byte was completely killing
performance.

Instead, always copy 16 bytes from the framebuffer onto the stack and
operate on that. Also fix the Pixel class to not operate on bytes to
avoid slower byte by byte loads that the compiler can't optimize away.

Done-with: Piotr Srebrny <piotr.srebrny@qt.io>

Change-Id: I96a58f38ed89273b73daa89ab95eca7fea13f3dd
Reviewed-by: Rafael Roquetto <rafael.roquetto@qt.io>