Provide QRhiTexture from QVideoFrame on windows backend
This patch implements a converion of OpenGL and D3D11 texture handles
to QRhiTexture in when creating the QVideoFrame on the Windows backend.

Additional, the patch explicitly states that the only supported fromats
are RGB8. D3D9 can use also RGB5 and RGB10A2. However, the RHI side
does not support RGB5. RGB10A2 could be added if required.

Change-Id: I4fa46988e7402926c99159ca5fbd597f1821eab0
Reviewed-by: Lars Knoll <lars.knoll@qt.io>