Encapsulate setting of encoder options for FFmpeg
FFmpeg doesn't have a unified way to specify encoder options, and
settings that work for one encoder could cause other encoders to
error out. So add some infrastructure that sets the options depending
on the encoder being used.

For now add support for libx264, libx265, libvpx and the videotoolbox
encoders. This can be easily extended to other encoders in the future.

Change-Id: I0010683096f1a8912a770ba254a0dd25513af986
Reviewed-by: Rafael Roquetto <rafael.roquetto@qt.io>