Link static ffmpeg library to ffmpeg multimedia plugin
This patch enables linking of the static FFmpeg library.

When linking the static library we have to provide the library
dependencies. These are found in the pkgconfig files generated during
the library build and are parsed with
the __ffmpeg_internal_set_dependencies function. The function adds
dependencies listed in lines starting with Libs and Libs.private.
While dependencies from Libs.private should suffice this is not
the case on Windows where dependencies are listed in the Libs line.

The -Wl,--exclude-libs=libavcodec flag fixes linking on Linux which
otherwise fails with relocation problems on ff_pw_9 symbol in
libavcodec. The flag prevents export of the libavcodec symbols, thus
linker can optimize access to the static variables (including ff_pw_9)
which otherwise are bound to the shared library GOT.

The patch fixes the order of FFmpeg libraries linked to FFmpeg plugin
to correctly display the dependencies between FFmpeg libraries.

Change-Id: I119d975fcff2b610280e27478ad186b4ffbc2b25
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>