darwin: Don't specify source format hint for audio recording
We were initializing the audio writer input with a format hint equal
to the audio capture device's active format. This is wrong as we
can't ensure that the captured audio buffers will be in that same
format. This caused failure in recording because the captured audio
buffers had the kAudioFormatFlagIsNonInterleaved set, while the format
hint didn't, causing the internal calls to the AudioConverter to fail
with errors.
To fix, do not specify a format hint but instead setup the audio
settings dictionary with the necessary keys. Force a default
value for AVNumberOfChannelsKey since it is now required, and for
recording with more than 2 channels, provide a supported channel
layout.

Fixes: QTBUG-98262
Done-with: Vladimir Belyavsky <belyavskyv@gmail.com>
Change-Id: I537d82e729c91a66d85aaa61c563a1ecd1a91af3
Reviewed-by: Vladimir Belyavsky <belyavskyv@gmail.com>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 980c145069c589085b4dee4b4c2b69e3b1a0adc3)
Reviewed-by: Doris Verria <doris.verria@qt.io>