Refactor QQnxMediaPlayer
Simplify (and fix) QQnxMediaPlayer by not having it store any state (to the
possible extent, i.e., some properties are cached because there is no
way to derive their value from the MMR API). The goal is to simplify the
logic, and thus, the code, by simply relying on the current state
reported by the MMR backend. This removes the problem when the
"recorded" state goes out of sync in relation to the backend.

This refactoring thus comprises:
    - removal and clean up of several member functions and variables
    - smooth video scrubbing by "batching" "set position" operations
    - fix progress reporting/current position
    - fix capacity reporting
    - fix scrubbing when paused
    - fix stop behavior.
    - remove obsolete autoPlay functionality
    - fix volume level setting
    - cleanup dead/messy code
    - fix setting the playback rate
    - relying on QPlatformMediaPlayer for caching state

Change-Id: I7acccf4f9cd7b9704f51faeee54c88d7fad1fe51
Reviewed-by: James McDonnell <jmcdonnell@blackberry.com>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>