Add support for proper HDR rendering
This is a larger change adding various pieces required to properly
playback HEVC encoded HDR videos.

Those videos are usually encoded in the BT2020 color space, but have
a non linear transfer function (perceptual quantizer or hybrid log gamma)
that is required to properly map the encoded video data to linear space.

In addition, the resulting pixels need to be tonemapped back to the supported
range of the display, something we do using the algorithm defined in
BT2390-6.

Finally, we need to acquire a HDR enabled surface from RHI to be able to
properly render the final result to screen.

This is tested on macOS, where we now get very good rendering of HDR videos.
Some more work is required to implement support for HLG transfer functions,
rendering properly on SDR surfaces (where we need to apply an inverse gamma
transform) and to support more HDR pixel formats.

Change-Id: Ida180260622136ee1f9dc9ae19781c3529424ebc
Reviewed-by: Rafael Roquetto <rafael.roquetto@qt.io>