Expose the color range being used in the video data
Add QVideoFrameFormat::colorRange, and expose the color range being
used in the video data.

Deprecate the full range color space enum values. This is much better
exposed through a separate property. As a drive by fix and clarify the
naming of the JPEG color space by renaming it to AdobeRgb.

Change-Id: Ifcdadcca41fdeade7d5ff2b80f59c135518d5d39
Reviewed-by: Lars Knoll <lars.knoll@qt.io>