QNX: QnxAudioSink - fix requested buffer size API
Enable setting and getting the buffer size (aka the fragment size) on
QQnxAudioSink.

It is important to note that the backend may not comply with the
requested buffer size, and may approximate it to the nearest compatible
value. It is thus possible that a call to bufferSize() returns a
different value than the one passed to a previous call to
setBufferSize().

The default buffer size value is set to be the maximum value reported by
the backend.

Change-Id: Ie4728d4effbd71fd6e8d23db020f46cd3eec3510
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: James McDonnell <jmcdonnell@blackberry.com>