QQnxAudioSink: add changeState() API
Replace setState and setError with an "atomic" changeState(State, Error)
API which forces the caller to _always_ specify the desired state and
error as a pair. This makes it much easier for the reader to visualize
the internal state transitions at any given point in code, and is less
error-prone.

Change-Id: Iab1057808e3e48f165e4fc321de350ff380d1437
Reviewed-by: Rafael Roquetto <rafael.roquetto@qt.io>
Reviewed-by: James McDonnell <jmcdonnell@blackberry.com>