Move the low level audio integration back into Qt Multimedia
Those integrations are selected at compile time and not dependent
on native vs ffmpeg backend. Because of that we can move them
back into Qt Multimedia directly.

This will simplify and improve a couple of things:

* We don't need to build the audio code twice (once per plugins)
* Low level audio is always available even if you choose not to
  ship a large backend for audio/video decoding and recording
* We can use Qt Multimedia as a backend for other modules requiring
  audio only without bloating things there (e.g. for Qt Speech)
* We can directly integrate support for 3D audio and headphone
  spatialization into Qt Multimedia without requiring an additional
  module for that.

To do so, we need to remove the camera handling from
QPlatformMediaDevices. This is only partially done in this
commit (by adding a QPlatformVideoDevices class), and will
be further cleaned up in followup commits.

Change-Id: Ib4a33d7255faaf26dd825527786eae10ed283490
Reviewed-by: Lars Knoll <lars.knoll@qt.io>