Restart pulling data if QIODevice signals readyRead
This patch improves handling of the QIODevice protocol by QAudioSink
in the pull mode. QAudioSink will restart pulling data when it receives
the readyRead signal. If the QIODevice source is atEnd(), QAudioSink
will no longer emit UnderrunError, but will switch to IdleState without
error. It will first check if the device is open and ask for
bytesAvailable before performing read.

Change-Id: Idd1d18419f7b77d2df2d76e7b1892194d652d63a
Reviewed-by: Lars Knoll <lars.knoll@qt.io>