QNX: QQnxAudioSource - allow capture device to be specified
Allow QQnxAudioSource to be passed a QAudioDevice so that it can open
an arbitrary capture device (analogous to the current behavior in
QQnxAudioSink), rather than always the preferred one.

The device is opened with QQnxAudioUtils::openPcmDevice, which lets a
device id to be specified. Consequently, the type of m_pcmHandle has
been changed to QnxAudioUtils::HandleUniquePtr.

Change-Id: I3997eb805346a3f3c94d8ee85cf3d1010bb5c5f6
Reviewed-by: James McDonnell <jmcdonnell@blackberry.com>