QQnxAudioSource: add changeState() API
Replace setState and setError with an "atomic" changeState(State, Error)
API which forces the caller to _always_ specify the desired state and
error as a pair. This makes it much easier for the reader to visualize
the internal state transitions at any given point in code, and is less
error-prone.

Change-Id: I3b76817b5b0efb023551a53c0f9970a27bed6c6c
Reviewed-by: James McDonnell <jmcdonnell@blackberry.com>