Add Eigen to 3rdparty
This is required for the spatial audio support and used by
the resonance audio code base.

As Eigen is rather large, this is trimmed to what we need
for resonance-audio.

This also reduces the licenses to MPL2 only (exept for one file
that is included but not used which is Apache licensed), as code
licensed differently is not included and used anymore.

Change-Id: I12574b38812c5702fbade5fd318df74b6023d57c
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Kai Koehne <kai.koehne@qt.io>